+++ app/src/test/java/woowacourse/shopping/main/MainActivityTest.kt
@@ -1,18 +1,19 @@
-package woowacourse.shopping
+package woowacourse.shopping.main
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.ViewModelProvider
-import com.google.common.truth.Truth
+import com.google.common.truth.Truth.assertThat
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
-import woowacourse.shopping.ui.cart.CartActivity
-import woowacourse.shopping.ui.cart.CartViewModel
+import woowacourse.shopping.ui.MainActivity
+import woowacourse.shopping.ui.MainViewModel
 
 @RunWith(RobolectricTestRunner::class)
-class CartActivityTest {
+class MainActivityTest {
+
     @get:Rule
     var instantTaskExecutorRule = InstantTaskExecutorRule()
 
@@ -20,24 +21,24 @@ class CartActivityTest {
     fun `Activity 실행 테스트`() {
         // given
         val activity = Robolectric
-            .buildActivity(CartActivity::class.java)
+            .buildActivity(MainActivity::class.java)
             .create()
             .get()
 
         // then
-        Truth.assertThat(activity).isNotNull()
+        assertThat(activity).isNotNull()
     }
 
     @Test
     fun `ViewModel 주입 테스트`() {
         // given
         val activity = Robolectric
-            .buildActivity(CartActivity::class.java)
+            .buildActivity(MainActivity::class.java)
             .create()
             .get()
-        val viewModel = ViewModelProvider(activity)[CartViewModel::class.java]
+        val viewModel = ViewModelProvider(activity)[MainViewModel::class.java]
 
         // then
-        Truth.assertThat(viewModel).isNotNull()
+        assertThat(viewModel).isNotNull()
     }
 }
